Added FrameTimer class to control the game loop framerate.
This FrameTimer class limits the framerate of the game loop by using C++11's `<chrono>` library to measure the time between frames and ensure that the game runs at a consistent framerate passed in the constructor.

The Tick method returns a true if a frame has exceeded the specified frame duration.